Rename pip_install_incremental to pip_parse - Rename helpers and executables to match this new convention. - Make readme more readable. - Add integration test for the example.
