+++ examples/pip_parse/WORKSPACE
@@ -1,17 +1,16 @@
 workspace(name = "example_repo")
 
-load("@bazel_tools//tools/build_defs/repo:git.bzl", "git_repository")
+load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
 
-# TODO(hrfuller) change this to a released http_archive when a release exists.
-git_repository(
+http_archive(
     name = "rules_python",
-    remote = "https://github.com/bazelbuild/rules_python.git",
-    branch = "master"
+    url = "https://github.com/bazelbuild/rules_python/releases/download/0.1.0/rules_python-0.1.0.tar.gz",
+    sha256 = "b6d46438523a3ec0f3cead544190ee13223a52f6a6765a29eae7b7cc24cc83a0",
 )
 
-load("@rules_python//python:pip.bzl", "pip_install_incremental")
+load("@rules_python//python:pip.bzl", "pip_parse")
 
-pip_install_incremental(
+pip_parse(
     # (Optional) You can provide extra parameters to pip.
     # Here, make pip output verbose (this is usable with `quiet = False`).
     #extra_pip_args = ["-v"],
@@ -34,7 +33,7 @@ pip_install_incremental(
     requirements_lock = "//:requirements_lock.txt",
 )
 
-load("@pip_incremental//:requirements.bzl", "install_deps")
+load("@pip_parsed_deps//:requirements.bzl", "install_deps")
 
 # Initialize repositories for all packages in requirements_lock.txt.
 install_deps()
