+++ python/pip_install/BUILD
@@ -5,6 +5,7 @@ filegroup(
     srcs = glob(["*.bzl"]) + [
         "BUILD",
         "//python/pip_install/extract_wheels:distribution",
+        "//python/pip_install/parse_requirements_to_bzl:distribution",
     ],
     visibility = ["//:__pkg__"],
 )
