--- python/pip_install/create_incremental_repo/BUILD
@@ -1,21 +1,0 @@
-load("@rules_python//python:defs.bzl", "py_library", "py_test")
-load("//python/pip_install:repositories.bzl", "requirement")
-
-py_library(
-    name = "lib",
-    srcs = ["__init__.py"],
-    deps = [requirement("pip")]
-)
-
-py_test(
-    name = "test_create_incremental_repo",
-    size = "small",
-    srcs = [
-        "test_create_incremental_repo.py",
-    ],
-    tags = ["unit"],
-    deps = [
-        ":lib",
-        "//python/pip_install/extract_wheels/lib"
-    ],
-)
