--- python/pip_install/create_incremental_repo/extract_single_wheel/__main__.py
@@ -1,4 +1,0 @@
-from python.pip_install.create_incremental_repo.extract_single_wheel import main
-
-if __name__ == "__main__":
-    main()
