+++ python/pip_install/parse_requirements_to_bzl/extract_single_wheel/BUILD
@@ -1,0 +1,8 @@
+filegroup(
+    name = "distribution",
+    srcs = glob(
+        ["*"],
+        exclude = ["*_test.py"],
+    ),
+    visibility = ["//python/pip_install:__subpackages__"],
+)
